accepted-submitted, fixed the integer overflow error in case 13 since case 13 exceeds the maximum value that can be represented by a signed 32-bit integer (int)

+++ easy/2525.cpp
@@ -6,7 +6,7 @@ class Solution {
 public:
     std::string categorizeBox(int length, int width, int height, int mass) {
     std::string boxCategory1, boxCategory2, finalBoxCategory;
-    int volume = length * width * height;
+    long long int volume = static_cast<long long int>(length) * width * height;
 
     if((length >= std::pow(10,4)) || (width >= std::pow(10,4)) || (height>= std::pow(10,4)) || (mass >= std::pow(10,4)) || (volume >= std::pow(10,9))){
         boxCategory1 = "Bulky";
